Change package name to lowercase
