--- src/main/java/SunderRay/commands/Command.java
@@ -1,5 +1,0 @@
-package SunderRay.commands;
-
-public abstract class Command {
-    public abstract String execute();
-}
