--- src/main/java/SunderRay/commands/CommandWord.java
@@ -1,5 +1,0 @@
-package SunderRay.commands;
-
-public enum CommandWord {
-    LIST, MARK, UNMARK, DELETE, TODO, DEADLINE, EVENT, BYE;
-}
