--- src/main/java/SunderRay/commands/InvalidCommand.java
@@ -1,14 +1,0 @@
-package SunderRay.commands;
-
-public class InvalidCommand extends Command {
-    private final String message;
-
-    public InvalidCommand(String message) {
-        this.message = message;
-    }
-
-    @Override
-    public String execute() {
-        return message;
-    }
-}
