--- src/main/java/SunderRay/commands/MarkCommand.java
@@ -1,26 +1,0 @@
-package SunderRay.commands;
-
-import SunderRay.data.messages.InfoMsg;
-import SunderRay.tasklist.TaskList;
-import SunderRay.tasks.Task;
-
-public class MarkCommand extends Command {
-    private final TaskList taskList;
-    private final int taskId;
-    private final boolean isDone;
-
-    public MarkCommand(TaskList taskList, int taskId, boolean isDone) {
-        this.taskList = taskList;
-        this.taskId = taskId;
-        this.isDone = isDone;
-    }
-
-    @Override
-    public String execute() {
-        Task task = taskList.markTask(taskId, isDone);
-        return String.format(
-                "%s%n\t%s",
-                String.format(InfoMsg.MARK_TASK, isDone ? "mark" : "unmark"),
-                task);
-    }
-}
