--- src/main/java/SunderRay/data/icons/TaskIcon.java
@@ -1,18 +1,0 @@
-package SunderRay.data.icons;
-
-public enum TaskIcon {
-    TODO("T"),
-    DEADLINE("D"),
-    EVENT("E");
-
-    private final String type;
-
-    TaskIcon(String type) {
-        this.type = type;
-    }
-
-    @Override
-    public String toString() {
-        return this.type;
-    }
-}
