--- src/main/java/SunderRay/tasks/ToDo.java
@@ -1,14 +1,0 @@
-package SunderRay.tasks;
-
-import SunderRay.data.icons.TaskIcon;
-
-public class ToDo extends Task {
-    public ToDo(String description) {
-        super(description);
-    }
-
-    @Override
-    protected String getTaskIcon() {
-        return TaskIcon.TODO.toString();
-    }
-}
