+++ src/main/java/sunderray/commands/Command.java
@@ -1,0 +1,5 @@
+package sunderray.commands;
+
+public abstract class Command {
+    public abstract String execute();
+}
