+++ src/main/java/sunderray/commands/CommandWord.java
@@ -1,0 +1,5 @@
+package sunderray.commands;
+
+public enum CommandWord {
+    LIST, MARK, UNMARK, DELETE, TODO, DEADLINE, EVENT, BYE;
+}
