+++ src/main/java/sunderray/commands/ExitCommand.java
@@ -1,0 +1,14 @@
+package sunderray.commands;
+
+import sunderray.data.messages.InfoMsg;
+
+public class ExitCommand extends Command {
+    @Override
+    public String execute() {
+        return InfoMsg.END;
+    }
+
+    public static boolean isExit(Command command) {
+        return command instanceof ExitCommand;
+    }
+}
