+++ src/main/java/sunderray/commands/ListCommand.java
@@ -1,0 +1,21 @@
+package sunderray.commands;
+
+import sunderray.data.messages.InfoMsg;
+import sunderray.tasklist.TaskList;
+
+public class ListCommand extends Command {
+    private final TaskList taskList;
+
+    public ListCommand(TaskList taskList) {
+        this.taskList = taskList;
+    }
+
+    @Override
+    public String execute() {
+        if (taskList.getNumTasks() == 0) {
+            return InfoMsg.NO_TASKS;
+        }
+
+        return String.format("%s%n%s", InfoMsg.LIST_TASKS, taskList.toListDisplay());
+    }
+}
