+++ src/main/java/sunderray/data/icons/TaskIcon.java
@@ -1,0 +1,18 @@
+package sunderray.data.icons;
+
+public enum TaskIcon {
+    TODO("T"),
+    DEADLINE("D"),
+    EVENT("E");
+
+    private final String type;
+
+    TaskIcon(String type) {
+        this.type = type;
+    }
+
+    @Override
+    public String toString() {
+        return this.type;
+    }
+}
